fix: independent chat window has no data
